pep8 compatibility

+++ interfaces/default/js/couchpotato.js
@@ -1,43 +1,43 @@
-var profiles = $('<select>');
+var profiles = $('<select>')
 $(document).ready(function() {
     $(window).trigger('hashchange')
     getMovieList()
     getNotificationList()
     getHistory()
     $('#searchform').submit(function(e) {
-        e.preventDefault();
-        var search = $('#search_movie_name').val();
-        if (search) searchMovie(search);
-    });
+        e.preventDefault()
+        var search = $('#moviename').val()
+        if (search) searchMovie(search)
+    })
     $.get(WEBDIR + 'couchpotato/GetProfiles', function(data) {
         if (data == null) return
         $.each(data.list, function(i, item) {
-            if (!item.hide) profiles.append($('<option>').val(item.id).text(item.label));
-        });
-    });
-});
+            if (!item.hide) profiles.append($('<option>').val(item.id).text(item.label))
+        })
+    })
+})
 
 function getMovieList() {
-    var wanted = $('#wanted-grid').empty();
-    $('.spinner').show();
+    var wanted = $('#wanted-grid').empty()
+    $('.spinner').show()
     $.getJSON(WEBDIR + 'couchpotato/GetMovieList', function (result) {
-        $('.spinner').hide();
+        $('.spinner').hide()
         if (result == null || result.total == 0) {
             wanted.append($('<li>').html('No wanted movies found'))
             return
         }
         $.each(result.movies, function(i, movie) {
             var link = $('<a>').attr('href', '#').click(function(e) {
-                e.preventDefault();
-                showMovie(movie);
-            });
+                e.preventDefault()
+                showMovie(movie)
+            })
             var src = WEBDIR + 'couchpotato/GetImage?w=100&h=150&url=' + movie.library.info.images.poster[0]
-            link.append($('<img>').attr('src', src).addClass('thumbnail'));
-            var title = shortenText(movie.library.info.original_title, 12);
-            link.append($('<h6>').addClass('movie-title').html(title));
-            wanted.append($('<li>').attr('id', movie.id).append(link));
-        });
-    });
+            link.append($('<img>').attr('src', src).addClass('thumbnail'))
+            var title = shortenText(movie.library.info.original_title, 12)
+            link.append($('<h6>').addClass('movie-title').html(title))
+            wanted.append($('<li>').attr('id', movie.id).append(link))
+        })
+    })
 }
 function showMovie(movie) {
     if (movie.library) {
@@ -48,54 +48,58 @@ function showMovie(movie) {
         var info = movie
         var plot = movie.plot
         var year = movie.year
-    };
-    var src = 'holder.js/154x231/text:No artwork';
-    if (info.images.poster) {
+    }
+    var src = 'holder.js/154x231/text:No artwork'
+    if (info.images.poster && info.images.poster[0]) {
         src = WEBDIR + 'couchpotato/GetImage?w=154&h=231&url=' + info.images.poster[0]
     }
-    var modalImg = $('<img>').attr('src', src).addClass('thumbnail pull-left');
+    var modalImg = $('<img>').attr('src', src).addClass('thumbnail pull-left')
 
-    var modalInfo = $('<div>').addClass('modal-movieinfo');
+    var modalInfo = $('<div>').addClass('modal-movieinfo')
     if (info.runtime) {
-        modalInfo.append($('<p>').html('<b>Runtime:</b> ' + parseSec(info.runtime)));
+        modalInfo.append($('<p>').html('<b>Runtime:</b> ' + parseSec(info.runtime)))
     }
-    modalInfo.append($('<p>').html('<b>Plot:</b> ' + plot));
+    modalInfo.append($('<p>').html('<b>Plot:</b> ' + plot))
     if (info.directors) {
-        modalInfo.append($('<p>').html('<b>Director:</b> ' + info.directors));
+        modalInfo.append($('<p>').html('<b>Director:</b> ' + info.directors))
     }
     if (info.genres) {
-        modalInfo.append($('<p>').html('<b>Genre:</b> ' + info.genres));
+        modalInfo.append($('<p>').html('<b>Genre:</b> ' + info.genres))
     }
+
     if (info.rating && info.rating.imdb) {
-        var rating = $('<div>').raty({
-            readOnly: true,
-            path: WEBDIR+'img',
-            score: (info.rating.imdb[0] / 2),
-        })
-        modalInfo.append(rating);
+        modalInfo.append(
+            $('<div>').raty({
+                readOnly: true,
+                path: WEBDIR+'img',
+                score: (info.rating.imdb[0] / 2),
+            })
+        )
     }
 
     profiles.unbind()
-    var title = info.original_title
+    var title = info.original_title + ' ('+year+')'
     if (movie.library) {
         profiles.change(function() {
-            editMovie(movie.id, $(this).val());
+            editMovie(movie.id, $(this).val())
         }).val(movie.profile_id)
         var modalButtons = {
             'Delete' : function() {
-                deleteMovie(movie.id, title);
-                hideModal();
+                if (confirm('Do you want to delete: ' + title)) {
+                    deleteMovie(movie.id, title)
+                    hideModal()
+                }
             },
             'Refresh' : function() {
-                refreshMovie(movie.id, title);
-                hideModal();
+                refreshMovie(movie.id, title)
+                hideModal()
             }
         }
     } else {
         var modalButtons = {
             'Add' : function() {
-                addMovie(movie.imdb, profiles.val(), title);
-                hideModal();
+                addMovie(movie.imdb, profiles.val(), title)
+                hideModal()
             }
         }
     }
@@ -104,11 +108,12 @@ function showMovie(movie) {
             'IMDb' : function() {
                 window.open('http://www.imdb.com/title/'+info.imdb,'IMDb')
             }
-        });
+        })
     }
     modalInfo.append(profiles)
     var modalBody = $('<div>').append(modalImg, modalInfo)
-    showModal(title + ' ('+year+')',  modalBody, modalButtons);
+    showModal(title, modalBody, modalButtons)
+    Holder.run()
 }
 function addMovie(movieid, profile, title) {
     $.getJSON(WEBDIR + 'couchpotato/AddMovie', {
@@ -116,12 +121,14 @@ function addMovie(movieid, profile, title) {
         profile: profile,
         title: encodeURIComponent(title)
     }, function (result) {
-        notify('CouchPotato', title + ' successfully added!', 'info');
+        console.log(result)
+        if (result == null || result.success != true) return
+        notify('CouchPotato', title + ' successfully added!', 'info')
         setTimeout(function() {
             getMovieList()
             $('a[href=#wanted]').tab('show')
-        }, 500);
-    });
+        }, 500)
+    })
 }
 function editMovie(id, profile) {
     $.getJSON(WEBDIR + 'couchpotato/EditMovie', {
@@ -129,68 +136,68 @@ function editMovie(id, profile) {
         profile: profile
     }, function (result) {
         if (result.success) {
-            notify('CouchPotato', 'Profile changed', 'info');
-            getMovieList();
+            notify('CouchPotato', 'Profile changed', 'info')
+            getMovieList()
         } else {
-            notify('CouchPotato', 'An error occured.', 'error');
+            notify('CouchPotato', 'An error occured.', 'error')
         }
-    });
+    })
 }
 function deleteMovie(id, name) {
     $.getJSON(WEBDIR + 'couchpotato/DeleteMovie', {id: id}, function (result) {
         if (result.success) {
-            $('#' + id).fadeOut();
-            notify('CouchPotato', name + ' successfully deleted!', 'info');
+            $('#' + id).fadeOut()
+            notify('CouchPotato', name + ' successfully deleted!', 'info')
         } else {
-            notify('CouchPotato', 'An error occured.', 'error');
+            notify('CouchPotato', 'An error occured.', 'error')
         }
-    });
+    })
 }
 function refreshMovie(id, name) {
     $.getJSON(WEBDIR + 'couchpotato/RefreshMovie', {id: id}, function (result) {
         if (result.success) {
-            notify('CouchPotato', 'Refreshing: ' + name, 'info');
+            notify('CouchPotato', 'Refreshing: ' + name, 'info')
         } else {
-            notify('CouchPotato', 'An error occured.', 'error');
+            notify('CouchPotato', 'An error occured.', 'error')
         }
-    });
+    })
 }
 function searchMovie(q) {
     var grid = $('#result-grid').empty()
     $('a[href=#result]').tab('show')
-    $('#searchspinner').show()
+    $('.spinner').show()
     $.getJSON(WEBDIR + 'couchpotato/SearchMovie', {
         q: encodeURIComponent(q)
     }, function (result) {
-        $('.spinner').hide();
+        $('.spinner').hide()
         $.each(result.movies, function(i, movie) {
             var link = $('<a>').attr('href', '#').click(function(e) {
-                e.preventDefault();
-                showMovie(movie);
-            });
-            var src = 'holder.js/100x150/text:No artwork';
+                e.preventDefault()
+                showMovie(movie)
+            })
+            var src = 'holder.js/100x150/text:No artwork'
             if (movie.images.poster[0]) {
                 src = WEBDIR + 'couchpotato/GetImage?w=100&h=150&url=' + movie.images.poster[0]
             }
-            link.append($('<img>').attr('src', src).addClass('thumbnail'));
-            var title = shortenText(movie.original_title, 12);
-            link.append($('<h6>').addClass('movie-title').html(title));
-            grid.append($('<li>').attr('id', movie.id).append(link));
+            link.append($('<img>').attr('src', src).addClass('thumbnail'))
+            var title = shortenText(movie.original_title, 12)
+            link.append($('<h6>').addClass('movie-title').html(title))
+            grid.append($('<li>').attr('id', movie.id).append(link))
         })
-        Holder.run();
+        Holder.run()
     })
 }
 function getNotificationList() {
     $.getJSON(WEBDIR + 'couchpotato/GetNotificationList', function (result) {
-        if (result == null) return;
+        if (result == null) return
         $.each(result.notifications, function(i, item) {
-            if (!item.read) notify('Notifications', item.message, 'info');
-        });
-    });
+            if (!item.read) notify('Notifications', item.message, 'info')
+        })
+    })
 }
 function getHistory() {
     $.getJSON(WEBDIR + 'couchpotato/GetNotificationList', function (result) {
-        if (result == null) return;
+        if (result == null) return
         $.each(result.notifications, function(i, item) {
             $('#history-grid').prepend(
                 $('<tr>').append(
@@ -198,6 +205,6 @@ function getHistory() {
                     $('<td>').text(item.message)
                 )
             )
-        });
-    });
+        })
+    })
 }
